third_party/blink/web_tests/fast/dom/custom/lifecycle-created-createElement-recursion-expected.txt

CONSOLE WARNING: line 8: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE ERROR: line 8: Uncaught RangeError: Maximum call stack size exceeded.
PASS unless crash
PASS successfullyParsed is true

TEST COMPLETE
